+++ Sources/CreateProject/UserChoice.swift
@@ -1,35 +1,5 @@
 import Foundation
 
-extension ProcessInfo {
-    func string(for argument: UserChoice.Argument) -> String? {
-        let arguments = ProcessInfo.processInfo.arguments
-
-        guard let firstIndex = arguments.firstIndex(of: argument.rawValue) else {
-            return nil
-        }
-
-        guard arguments.indices.contains(firstIndex + 1) else {
-            return nil
-        }
-
-        return arguments[firstIndex + 1]
-    }
-
-    func bool(for argument: UserChoice.Argument) -> Bool? {
-        let arguments = ProcessInfo.processInfo.arguments
-
-        guard let firstIndex = arguments.firstIndex(of: argument.rawValue) else {
-            return nil
-        }
-
-        guard arguments.indices.contains(firstIndex + 1) else {
-            return nil
-        }
-
-        return Bool(arguments[firstIndex + 1])
-    }
-}
-
 enum UserChoice {
     enum Error: Swift.Error, LocalizedError {
         case missingUserChoice
@@ -39,38 +9,6 @@ enum UserChoice {
             switch self {
                 case .missingUserChoice: "Recieved no input from user"
                 case .invalidInputForBoolean: "Invalid input expect \"y\" or \"n\""
-            }
-        }
-    }
-
-    enum Argument: String, CaseIterable {
-        case help = "--help"
-        case noColor = "--noColor"
-        case projectName = "--projectName"
-        case executableName = "--executableName"
-        case projectPath = "--projectPath"
-        case godotPath = "--godot"
-        case version = "--version"
-
-        var description: String {
-            switch self {
-                case .help: "show help info"
-                case .noColor: "disable colorized output"
-                case .version: "show current version"
-                case .projectName: "project name"
-                case .executableName: "executable name"
-                case .godotPath: "path to Godot binary"
-                case .projectPath: "path where Package.swift and other files will be saved"
-            }
-        }
-
-        var usage: String {
-            switch self {
-                case .help, .noColor, .version:
-                    rawValue
-                
-                case .projectName, .executableName: rawValue + " <name>"
-                case .projectPath, .godotPath: rawValue + " <path>"
             }
         }
     }
